refactor(pricing): decompõe pricing.ts em shared-tiers + DEFAULT_PRICING particionado (godfile sweep, #3501)
pricing.ts 1661→97; consts de tier → pricing/shared-tiers.ts; DEFAULT_PRICING (30 chaves)
dividido em 2 partes <800 (cada uma importa só os tier-consts que usa) + spread-merge no
barrel. DEFAULT_PRICING byte-idêntico (integridade provada); helpers mantidos no host.
